add test for write() method
